pyproject.toml
[project]
name = "sphinx-lumina-theme"
version = "1.8.4"
description = "A modern Sphinx documentation theme with Tailwind CSS and Alpine.js"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "sphinx>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "beautifulsoup4>=4.12",
    "myst-parser>=4.0",
    "sphinx-design>=0.6",
    "sphinx-copybutton>=0.5",
    "sphinxcontrib-mermaid>=1.0",
    "sphinx-autobuild>=2024.0",
    "ruff>=0.11",
    "djlint>=1.36",
    "pytest-playwright>=0.6",
]

[project.entry-points."sphinx.html_themes"]
lumina = "sphinx_lumina_theme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.hooks.custom]

[tool.hatch.build]
artifacts = ["src/sphinx_lumina_theme/theme/static/lumina.css", "src/sphinx_lumina_theme/theme/static/lumina.js"]

[tool.hatch.build.targets.wheel]
packages = ["src/sphinx_lumina_theme"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ["src/sphinx_lumina_theme/__init__.py:__version__"]
build_command = "python3 scripts/update_version.py {version}"
commit_parser = "conventional"
major_on_zero = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
parse_squash_commits = true
ignore_merge_commits = true

[tool.semantic_release.changelog]
insertion_flag = "<!-- releases -->"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
